requirements.txt
amqp==5.0.7
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
Django==3.2.9
django-filter==21.1
django-redis==5.1.0
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
kombu==5.2.2
MarkupSafe==2.0.1
openapi-codec==1.3.2
packaging==21.3
Pillow==8.4.0
prompt-toolkit==3.0.24
psycopg2-binary==2.9.2
pyparsing==3.0.6
python-decouple==3.5
pytz==2021.3
redis==3.5.3
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
transliterate==1.10.2
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
